+++ test/message_extraction/embedded_plural_text_after_test.dart
@@ -5,7 +5,7 @@
 library embedded_plural_text_after_test;
 
 import "failed_extraction_test.dart";
-import "package:unittest/unittest.dart";
+import "package:test/test.dart";
 
 main() {
   test("Expect failure because of embedded plural with text after it", () {