+++ test/message_extraction/message_extraction_no_deferred_test.dart
@@ -2,14 +2,13 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-
 /// A test for message extraction and code generation not using deferred
 /// loading for the generated code.
 library message_extraction_no_deferred_test;
 
 import 'message_extraction_test.dart' as mainTest;
 
-main(arguments) {
+main() {
   mainTest.useDeferredLoading = false;
-  mainTest.main(arguments);
+  mainTest.main();
 }