+++ test/message_extraction/really_fail_extraction_test.dart
@@ -5,7 +5,7 @@
 library really_fail_extraction_test;
 
 import "failed_extraction_test.dart";
-import "package:unittest/unittest.dart";
+import "package:test/test.dart";
 
 main() {
   test("Expect failure because warnings are errors", () {